+++ FlashCardsManager/src/core/FlashCard.java
@@ -1,5 +1,7 @@
 package core;
 
+import importExport.*;
+
 import java.io.IOException;
 import java.sql.SQLException;
