+++ FlashCardsManager/src/importExport/XMLFlashcardExchanger.java
@@ -1,4 +1,4 @@
-package export;
+package importExport;
 
 import java.io.*;
 import java.util.*;
